Fix/artifact errors

+++ slack-javascript-action/src/approval.helpers.ts
@@ -2,10 +2,10 @@ import { sendMessage, SlackMessage, updateMessage } from "./slack.client";
 import useBlocks from "./useBlocks";
 import { ConfigType } from "./useConfig";
 import fs from "fs";
-import * as artifact from "@actions/artifact";
+import { DefaultArtifactClient } from "@actions/artifact";
 import { threadReleaseNotes } from "./deploy.helpers";
 
-const FILE_NAME = "release-message-information.config";
+const FILE_NAME = "release-message-information";
 
 const askForApproval = async (config: ConfigType) => {
   const message = await sendMessage(
@@ -18,14 +18,15 @@ const askForApproval = async (config: ConfigType) => {
   });
 
   fs.writeFileSync(
-    "./release-message-information.config",
+    FILE_NAME,
     JSON.stringify({
       channel: config.channel,
       message_id: message.ts,
     }),
     { encoding: "utf8" }
   );
-  await artifact.create().uploadArtifact(FILE_NAME, [`./${FILE_NAME}`], ".");
+  const artifactClient = new DefaultArtifactClient();
+  await artifactClient.uploadArtifact(FILE_NAME, [`./${FILE_NAME}`], ".");
 
   return message.ts;
 };
@@ -41,22 +42,27 @@ const approvalWasGranted = async (
   );
 
   try {
-    const downloadResponse = await artifact
-      .create()
-      .downloadArtifact(FILE_NAME, ".");
-    const fileLocation = `${downloadResponse.downloadPath}/${FILE_NAME}`;
-    const content = fs.readFileSync(fileLocation, {
-      encoding: "utf8",
-    });
-    fs.unlink(fileLocation, () => {});
-    const { message_id, channel } = JSON.parse(content);
+    const artifactClient = new DefaultArtifactClient();
+    const { artifact } = await artifactClient.getArtifact(FILE_NAME);
+    if (artifact) {
+      const { downloadPath } = await artifactClient.downloadArtifact(
+        artifact.id
+      );
 
-    let message;
-    if (message_id) {
-      message = await updateMessage(channel, message_id, messageBlocks, "");
+      const fileLocation = `${downloadPath}/${FILE_NAME}`;
+      const content = fs.readFileSync(fileLocation, {
+        encoding: "utf8",
+      });
+      fs.unlink(fileLocation, () => {});
+      const { message_id, channel } = JSON.parse(content);
+
+      let message;
+      if (message_id) {
+        message = await updateMessage(channel, message_id, messageBlocks, "");
+      }
+
+      return message?.ts;
     }
-
-    return message?.ts;
   } catch (e) {
     console.info("Problem updating file:", e);
   }
